+++ generators/view/inputs/viewName.js
@@ -40,7 +40,7 @@
 		static set($that, $defaultFromMemory) {
 			logs.hintsMultiline($that, hints);
 
-			prompts[0].default = $defaultFromMemory ? $that.viewName : 'viewName';
+			prompts[0].default = $defaultFromMemory ? $that.viewName : $that.mainDirectory;
 
 			return $that.prompt(prompts).then($response => {
 				this.onSuccess($that, $response);