+++ generators/view/inputs/stateName.js
@@ -31,7 +31,7 @@
 		static set($that, $defaultFromMemory) {
 			logs.hintsMultiline($that, hints);
 
-			prompts[0].default = $defaultFromMemory ? $that.stateName : $that.viewNameCamel;
+			prompts[0].default = $defaultFromMemory ? $that.stateName : $that.viewNameDotted;
 
 			return $that.prompt(prompts).then($response => {
 				this.onSuccess($that, $response);